Add NoteList component and improve note handling
Added NoteList component to display list of notes in the NotesPage. Made changes in note handling where form data is now converted to an object, a unique timestamp ID is added to each new note, and this note is appended to the existing list of notes. These changes aimed to provide a more user-friendly note listing and effective note management system.